fix(notes): allow editing notes
